% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NVIpkg_skeleton.R
\name{create_NVIpkg_skeleton}
\alias{create_NVIpkg_skeleton}
\title{Create the package skeleton for NVIverse packages}
\usage{
create_NVIpkg_skeleton(
  pkg = stringi::stri_extract_last_words(usethis::proj_path()),
  pkg_path = usethis::proj_path(),
  license_keyword = "Apache-2.0"
)
}
\arguments{
\item{pkg}{[\code{character(1)}]\cr
The package name. Defaults to \code{stringi::stri_extract_last_words(usethis::proj_path())}.}

\item{pkg_path}{[\code{character(1)}]\cr
The path to the package directory. Defaults to \code{usethis::proj_path()}.}

\item{license_keyword}{[\code{character(1)}]\cr
The keyword for the package's license in accord with list of license keywords.
Defaults to "BSD_3_clause".}
}
\value{
None.
    Sets up the package directories and writes and modifies several files, see details.
}
\description{
Creates the package skeleton in agreement with the conventions
    developed for NVIverse packages. \code{create_NVIpkg_skeleton} should be
    run once after a GitHub repository has been synchronized with the the
    package directory.
}
\details{
\code{create_NVIpkg_skeleton} is a wrapper for several \code{usethis}
    -functions. It sets up the package directory with standard files and
    standard sub-directories. Modifies the \code{DESCRIPTION}, \code{.gitignore},
    and \code{.Rbuildignore} in agreement with standard dependencies for
    \code{NVIverse}.

    Standard values are input to \code{DESCRIPTION}. For modifying these values,
    you need to modify \code{\link{set_description_default}}.

    In addition \code{README.Rmd}-template, \code{CONTRIBUTING}, \code{CODE_OF_CONDUCT},
    and the vignette \code{Contribute_to_NVIpkg} are copied to the package
    directory and \code{develop.R} is copied to "./notes/".

    The function accepts a short list of the listed license keywords accepted
    at Cran in "./share/license/license.db" in R home. If you need any  other
    license than in the short list, submit an issue.
}
\examples{
\dontrun{
# Attach packages and set up with temporary directory
library(NVIpackager)
td <- tempdir()
if (!dir.exists(file.path(td, "NVItest"))) {
  dir.create(file.path(td, "NVItest"))
}
# Create package skeleton in temporary directory
create_NVIpkg_skeleton(pkg = "NVItest", pkg_path = file.path(td, "NVItest"))
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
