% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vignettes.R
\name{get_vignettes}
\alias{get_vignettes}
\title{Gets the title of the package vignettes}
\usage{
get_vignettes(path = file.path(usethis::proj_path(), "vignettes"))
}
\arguments{
\item{path}{[\code{character(1)}]\cr
directory with vignettes. Defaults to subdirectory "vignettes" in the package directory.}
}
\value{
character vector with vignette titles.
}
\description{
Gets the title of the vignettes in the package. The function is intended for listing the vignette titles in README.
}
\details{
The title is extracted from the vignette files in the package directory ./vignettes/. to make it possible to extract the information before the package is installed. This in contrast to \code{base::vignettes} that get the titles from the installed package.
}
\examples{
\dontrun{
# Used when are in the package directory
get_vignettes()
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
