% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NVIpackages}
\alias{NVIpackages}
\title{Data: Description of the NVIverse packages .}
\format{
A data frame with 3 variables:
\describe{
\item{Package}{name of the NVIverse package.}
\item{Status}{Whether the package is "Private" or "Public".}
\item{Description}{The short description of the package as given in the "Title" field in \code{DESCRIPTION}.}
}
}
\source{
\code{./data-raw/generate_NVIpackages.R} in package \code{NVIpackager}
}
\usage{
NVIpackages
}
\description{
A data frame with key information for the NVIverse packages. The information is used as input to \code{README.RMD} among others.
The code for preparing of the data frame is written in
\code{./data-raw/generate_NVIpackages.R} and the raw data can be edited there.
}
\keyword{datasets}
