% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_description_default.R
\name{set_description_default}
\alias{set_description_default}
\title{Set default Values for the \code{DESCRIPTION} file}
\usage{
set_description_default(
  pkg = stringi::stri_extract_last_words(usethis::proj_path()),
  license_keyword = "BSD_3_clause"
)
}
\arguments{
\item{pkg}{The package name.}

\item{license_keyword}{The keyword for the package's license in accord with list of license keywords.}
}
\value{
None. Writes the options to the option \code{usethis.description}.
}
\description{
Set default Values for the \code{DESCRIPTION} file. The values are saved
as the option \code{usethis.description}.
}
\details{
These options are used when creating the package skeleton in the script
\code{create_NVIpackage_skeleton.R}.\preformatted{There is a list of license keywords at
[GitHub](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/licensing-a-repository)
}
}
\examples{
\dontrun{
}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
