% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_contribute.R
\name{update_contribute}
\alias{update_contribute}
\title{Update the vignette "Contribute to NVIpkg.Rmd" and CONTRIBUTING.md}
\usage{
update_contribute(
  pkg = stringi::stri_extract_last_words(usethis::proj_path()),
  pkg_path = usethis::proj_path()
)
}
\arguments{
\item{pkg}{The package name.}

\item{pkg_path}{The path for the package}
}
\value{
None. Writes the vignette and the file CONTRIBUTING.md.
}
\description{
Update the vignette "Contribute to NVIpkg.Rmd" and CONTRIBUTING.md
from the template in NVI-package.
}
\details{
The template for CONTRIBUTING.md is found in NVIpackager. Any change in the file must be done in the template.
}
\examples{
\dontrun{
# Attach packages and set up with temporary directory
td <- tempdir()
library(NVIpackager)

pkg <-
 use_contribute(pkg = pkg)
}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
