% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_license.R
\name{update_license}
\alias{update_license}
\title{Update copyright year in the license}
\usage{
update_license(
  pkg = stringi::stri_extract_last_words(usethis::proj_path()),
  pkg_path = usethis::proj_path(),
  copyright_owner = "Norwegian Veterinary Institute"
)
}
\arguments{
\item{pkg}{[\code{character(1)}]\cr
The package name. Defaults to \code{stringi::stri_extract_last_words(usethis::proj_path())}.}

\item{pkg_path}{[\code{character(1)}]\cr
The path to the package directory. Defaults to \code{usethis::proj_path()}.}

\item{copyright_owner}{[\code{character(1)}]\cr
The copyright owner in the copyright statement. Defaults
    to "Norwegian Veterinary Institute".}
}
\value{
None. Writes the LICENSE and LICENSE.md files.
}
\description{
Update copyright year in the LICENSE and LICENSE.md files. The
    copyright years are given as a range from the first year
    to the current year.
}
\details{
The copyright years will only be updated for the
    given copyright owner This to avoid updating the copyright year for
    other copyright owners than the copyright owner.
}
\examples{
\dontrun{
# Used when are in the package directory
update_license()
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
