% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_logo.R
\name{update_logo}
\alias{update_logo}
\title{Update logo}
\usage{
update_logo(
  pkg = stringi::stri_extract_last_words(usethis::proj_path()),
  pkg_path = usethis::proj_path()
)
}
\arguments{
\item{pkg}{[\code{character}]\cr
The package name, defaults to \code{stringi::stri_extract_last_words(usethis::proj_path())}.}

\item{pkg_path}{[\code{character}]\cr
The path to the package directory, defaults to \code{usethis::proj_path()}.}
}
\value{
None. Copies the the logo file to the directory "./man/figures/.
}
\description{
Update the logo to be used in the README file.
}
\details{
If the package logo has been created, the logo is copied from
\code{NVIrpackages} to the package directory "./man/figures/. If the logo
already exists, the logo is overwritten.
}
\examples{
\dontrun{
# Attach packages and set up with temporary directory
library(NVIpackager)
update_logo()
}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
