% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_readme.R
\name{update_readme}
\alias{update_readme}
\title{Update \code{README.md}}
\usage{
update_readme(
  pkg = stringi::stri_extract_last_words(usethis::proj_path()),
  pkg_path = usethis::proj_path()
)
}
\arguments{
\item{pkg}{[\code{character}]\cr
The package name, defaults to \code{stringi::stri_extract_last_words(usethis::proj_path())}.}

\item{pkg_path}{[\code{character}]\cr
The path to the package directory, defaults to \code{usethis::proj_path()}.}
}
\value{
None. Writes the file \code{README.md}.
}
\description{
Update \code{README.md}
from \code{README.Rmd} and child templates in \code{NVIpackager}. If the file don't exist, it is created.
}
\details{
The child templates are found in \code{NVIpackager}. Any change in this
text must be done in the templates.
}
\examples{
\dontrun{
# Attach packages and set up with temporary directory
library(NVIpackager)
td <- tempdir()
if (!dir.exists(file.path(td, "NVItest"))) {
  dir.create(file.path(td, "NVItest"))
}

update_readme(pkg = "NVItest",
              pkg_path = file.path(td, "NVItest"))
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
