% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_reference_manual.R
\name{update_reference_manual}
\alias{update_reference_manual}
\title{Update reference manual}
\usage{
update_reference_manual(
  pkg = stringi::stri_extract_last_words(usethis::proj_path()),
  pkg_path = usethis::proj_path(),
  manual
)
}
\arguments{
\item{pkg}{[\code{character}]\cr
The package name, defaults to \code{stringi::stri_extract_last_words(usethis::proj_path())}.}

\item{pkg_path}{[\code{character}]\cr
The path to the package directory, defaults to \code{usethis::proj_path()}.}

\item{manual}{Can be c("include", "update", "remove")}
}
\value{
None. Creates or updates the PDF reference manual in directory "./vignettes/.
The reference manual is given the name pkgname.pdf.
}
\description{
Update the PDF reference manual.
}
\details{
If the package logo has been created, the logo is copied from
\code{NVIrpackages} to the package directory "./man/figures/. If the logo
already exists, the logo is overwritten.
}
\examples{
\dontrun{
# Update the reference manual if it already exists
library(NVIpackager)
update_reference_manual(manual = "update")
}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
