% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NVIpkg_skeleton.R
\name{create_NVIpkg_skeleton}
\alias{create_NVIpkg_skeleton}
\title{Create the package skeleton for NVIverse packages}
\usage{
create_NVIpkg_skeleton(
  pkg = stringi::stri_extract_last_words(usethis::proj_path()),
  pkg_path = usethis::proj_path()
)
}
\arguments{
\item{pkg}{The package name.}

\item{pkg_path}{The path for the package}
}
\value{
None. Writes the DESCRIPTION and set up various directories and updates
.gitignore, .Rbuildignore, README, CONTRIBUTING, etc.
}
\description{
Create the package skeleton for NVIverse packages. The function is
wrapper for several usethis-functions. In addition templates for standard
files are copied to the package directory.
}
\details{
The function should be run after a GitHub repository has been syncronized
with the the directory and the set_description_default has been run.
}
\examples{
\dontrun{
# Attach packages and set up with temporary directory
library(NVIpackager)
td <- tempdir()

}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
