% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NVIpkg_skeleton.R
\name{create_NVIpkg_skeleton}
\alias{create_NVIpkg_skeleton}
\title{Create the package skeleton for NVIverse packages}
\usage{
create_NVIpkg_skeleton(
  pkg = stringi::stri_extract_last_words(usethis::proj_path()),
  pkg_path = usethis::proj_path()
)
}
\arguments{
\item{pkg}{The package name.}

\item{pkg_path}{The path to the package directory.}
}
\value{
None.
Sets up the package directories and writes and modifies several files, see details.
}
\description{
Creates the package skeleton in agreement with the conventions
developed for NVIverse packages. \code{create_NVIpkg_skeleton} should be
run once after a GitHub repository has been synchronized with the the
package directory and the \code{set_description_default} has been run.
}
\details{
\code{create_NVIpkg_skeleton} is a wrapper for several \code{usethis}
-functions. It sets up the package directory with standard files and
standard sub-directories. Modifies the \code{DESCRIPTION}, \code{.gitignore},
and \code{.Rbuildignore} in agreement with standard dependencies for
NVIverse.\preformatted{In addition \code{README.Rmd}-template, \code{CONTRIBUTING}, \code{CODE_OF_CONDUCT},
and the vignette \code{Contribute_to_NVIpkg} are copied to the package
directory.
}
}
\examples{
\dontrun{
# Attach packages and set up with temporary directory
library(NVIpackager)
td <- tempdir()

}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
