% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_NVIpkg.R
\name{document_NVIpkg}
\alias{document_NVIpkg}
\title{Update styling and documentation of a package}
\usage{
document_NVIpkg(
  pkg,
  pkg_path,
  style = FALSE,
  contributing = FALSE,
  readme = FALSE,
  ...
)
}
\arguments{
\item{style}{{logical}, defaults to FALSE.}

\item{contributing}{{logical}, defaults to FALSE.}

\item{readme}{{logical}, defaults to FALSE.}
}
\value{
none. Updated help files for all functions and,
depending on argument input, can updated style,
\code{CONTRIBUTING}, and \code{README}
}
\description{
A wrapper around functions for styling,
making the help files and updated vignettes and README.
}
\details{
The help files for R-functions will always be generated.
If styling and generation of CONTRIBUTING. md and README
can be control led by input arguments.
}
\examples{
\dontrun{
# Reading from standard directory at NVI's network
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
