% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_NVIpkg.R
\name{install_NVIpkg}
\alias{install_NVIpkg}
\title{Install an NVIverse Package during the Development Phase}
\usage{
install_NVIpkg(
  pkg = stringi::stri_extract_last_words(usethis::proj_path()),
  pkg_path = usethis::proj_path(),
  lib = R.home(),
  rsource,
  username = "NorwegianVeterinaryInstitute",
  ...
)
}
\arguments{
\item{pkg}{[\code{character}]\cr
The package name, defaults to \code{stringi::stri_extract_last_words(usethis::proj_path())}.}

\item{pkg_path}{[\code{character}]\cr
The path to the package directory, defaults to \code{usethis::proj_path()}.}

\item{lib}{[\code{character}]\cr
The library directory where to install, defaults to \code{R.home()}.}

\item{rsource}{[\code{character}]\cr
One of c("github", "local").}

\item{username}{[\code{character}]\cr
The github username where the repository is found, defaults to
 "NorwegianVeterinaryInstitute".}

\item{\dots}{Other arguments to be passed to \code{remotes::install_github}
or \code{devtools::install}.}
}
\value{
None. Installs a package.
}
\description{
Installation of an NVIverse package from github or
    from local directory. Mainly intended for installation of a
    package during development to test new code in the package
    scripts .
}
\details{
When \code{rsource = "github"}, the default is to install
    the latest version in the main branch at the NorwegianVeterinaryInstitute
    package repository. During package development one will usually
    change the username to your own.

The repository can be changed by giving more arguments that
    will be passed to \code{install_github}, see
    \link[remotes]{install_github} for full description
    of the arguments \code{repo = }, \code{username = }, and
    \code{ref = }.

When \code{rsource = "local"}, it installs the package from a
    local copy of the package repository. It defaults to install
    the active branch. Use \code{rsource = "local"} to test new
    code during development.

For installing the latest released versions of NVIverse
    packages, use \link[remotes]{install_github} or
    \link[NVIbatch]{use_NVIverse}.
}
\examples{
\dontrun{
# Installs a package from local PC
# You need to be in a package directory at you PC for the code to work.
pkg <- stringi::stri_extract_last_words(usethis::proj_path())
pkg_path = usethis::proj_path()

NVIpackager::install_NVIpkg(pkg = pkg, pkg_path = pkg_path, rsource = "local")
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
