% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_NVIpkg.R
\name{install_NVIpkg}
\alias{install_NVIpkg}
\title{Install an NVIverse Package during the Development Phase}
\usage{
install_NVIpkg(
  pkg,
  pkg_path,
  lib = R.home(),
  source,
  username = "NorwegianVeterinaryInstitute",
  ...
)
}
\arguments{
\item{pkg}{[\code{character}]\cr
The package name, defaults to \code{stringi::stri_extract_last_words(usethis::proj_path())}.}

\item{pkg_path}{[\code{character}]\cr
The path to the package directory, defaults to \code{usethis::proj_path()}.}

\item{lib}{character giving the library directory where to install
the package. If missing, defaults to the first element of
\code{.libPaths()}.}

\item{source, }{character with one of c("github", "local") .}

\item{username, }{the github username to install from. Defaults to
"NorwegianVeterinaryInstitute".}

\item{\dots}{Other arguments to be passed to \code{remotes::install_github}
or \code{devtools::install}.}
}
\value{
none. Installs a package.
}
\description{
Installation of an NVIverse package from github or
from local directory. Mainly intended for installation of a
package during development to test new code in the package
scripts .
}
\details{
When \code{source = "github"}, the default is to install
the latest version in the main branch at the NorwegianVeterinaryInstitute
package repository. During package development one will usually
change the username to your own.\preformatted{The repository can be changed by giving more arguments that
will be passed to \code{install_github}, see
\code{\link[remotes:install_github]} for full description
of the arguments \code{repo =}, \code{username =}, and
\code{ref =}.

When \code{source = "local"}, it installs the package from a
local copy of the package repository. It defaults to install
the active branch. Use \code{source = "local"} to test new
code during development.

For installing the latest released versions of NVIverse
packages, use \code{\link[remotes:install_github]} or
\code\{\link[NVIbatch:use_NVIverse].
}
}
\examples{
\dontrun{
# Reading from standard directory at NVI's network
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
