% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_NVIpkg.R
\name{install_NVIpkg}
\alias{install_NVIpkg}
\title{Update styling and documentation of a package}
\usage{
install_NVIpkg(
  pkg,
  pkg_path,
  Rlibrary = R.home(),
  type,
  repo = "NorwegianVeterinaryInstitute",
  branch,
  ...
)
}
\arguments{
\item{Rlibrary, }{defaults to FALSE.}

\item{type, }{defaults to FALSE.}

\item{repo, }{defaults to FALSE.}

\item{branch, }{defaults to FALSE.}

\item{dots, }{defaults to FALSE.}
}
\value{
none. Updated help files for all functions and,
depending on argument input, can updated style,
\code{CONTRIBUTING}, and \code{README}
}
\description{
A wrapper around functions for styling,
making the help files and updated vignettes and README.
}
\details{
The help files for R-functions will always be generated.
If styling and generation of CONTRIBUTING. md and README
can be control led by input arguments.
}
\examples{
\dontrun{
# Reading from standard directory at NVI's network
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
