% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVIpackager-package.R
\docType{package}
\name{NVIpackager-package}
\alias{NVIpackager}
\alias{NVIpackager-package}
\title{NVIpackager: Tools to facilitate Development of NVIverse Packages}
\description{
Provides tools to facilitate development of NVIverse packages. You should use 'create_NVIpkg_skeleton' to make a package skeleton in accord with NVIverse standards. For further development and maintenance there are tools for updating documentation and installing development versions. The NVIpackager functions are to a large extent wrappers for devtools and usethis functions. In addition, the templates folder keeps templates for standard files like sections of README and CONTRIBUTING.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NorwegianVeterinaryInstitute/NVIpackager}
  \item Report bugs at \url{https://github.com/NorwegianVeterinaryInstitute/NVIpackager/issues}
}

}
\author{
\strong{Maintainer}: Petter Hopp \email{Petter.Hopp@vetinst.no} (\href{https://orcid.org/0000-0002-8695-0378}{ORCID})

Other contributors:
\itemize{
  \item Norwegian Veterinary Institute [copyright holder]
}

}
\keyword{internal}
