% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_description_default.R
\name{set_description_default}
\alias{set_description_default}
\title{Set default Values for the \code{DESCRIPTION} file}
\usage{
set_description_default(
  pkg = stringi::stri_extract_last_words(usethis::proj_path()),
  license_keyword = "BSD_3_clause"
)
}
\arguments{
\item{pkg}{[\code{character(1)}]\cr
The package name. Defaults to \code{stringi::stri_extract_last_words(usethis::proj_path())}.}

\item{license_keyword}{[\code{character(1)}]\cr
The keyword for the package's license in accord with list of license keywords.
    Defaults to "BSD_3_clause".}
}
\value{
None. Writes the options to the option \code{usethis.description}.
}
\description{
Set default Values for the \code{DESCRIPTION} file. The values
    are saved as the option \code{usethis.description}.
}
\details{
These options are used when creating the package skeleton using
    \code{\link{create_NVIpkg_skeleton}}.
    The following standard values are given:

\itemize{
  \item "Petter Hopp" with role c("aut", "cre")
  \item "Norwegian Veterinary Institute" with role = "cph"
  \item registertype (categories for locations and addresses)
  \item URL: "https://github.com/NorwegianVeterinaryInstitute/NVIpkg"
  \item BugReports: "https://github.com/NorwegianVeterinaryInstitute/NVIpkg/issues"
  \item License: license_keyword "+ file LICENSE",
  \item Language: "en-GB",
  \item Depends:= "R (>= 4.0.0)"
}
    The function accepts a short list of the listed license keywords accepted
    at Cran in "./share/license/license.db" in R home. If you need any  other
    license than in the short list, submit an issue.
}
\examples{
\dontrun{
library(NVIpackager)
set_description_default()
}


# set up standards for DESCRIPTION file ----
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
