% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_contributing.R
\name{update_contributing}
\alias{update_contributing}
\title{Update \code{CONTRIBUTING.md}}
\usage{
update_contributing(
  pkg = stringi::stri_extract_last_words(usethis::proj_path()),
  pkg_path = usethis::proj_path()
)
}
\arguments{
\item{pkg}{[\code{character}]\cr
The package name, defaults to \code{stringi::stri_extract_last_words(usethis::proj_path())}.}

\item{pkg_path}{[\code{character}]\cr
The path to the package directory, defaults to \code{usethis::proj_path()}.}
}
\value{
None. Writes the vignette "Contribute_to_NVIpkg.Rmd" and the file \code{CONTRIBUTING.md}.
}
\description{
Update \code{CONTRIBUTING.md} and the vignette "Contribute_to_NVIpkg.Rmd"
from the template in \code{NVIpackager}. If the files don't exist, they are created.
}
\details{
The template is found in \code{NVIpackager}. Any change in the text
must be done in the template.
}
\examples{
\dontrun{
# Attach packages and set up with temporary directory
library(NVIpackager)
td <- tempdir()
if (!dir.exists(file.path(td, "NVItest"))) {
  dir.create(file.path(td, "NVItest"))
}
if (!dir.exists(file.path(td, "NVItest", "vignettes"))) {
  dir.create(file.path(td, "NVItest", "vignettes"))
}

use_contributing(pkg = "NVItest",
                 pkg_path = file.path(td, "NVItest"))
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
