% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_news.R
\name{update_news}
\alias{update_news}
\title{Update NEWS}
\usage{
update_news(
  pkg = stringi::stri_extract_last_words(usethis::proj_path()),
  pkg_path = usethis::proj_path(),
  template = NULL
)
}
\arguments{
\item{pkg}{[\code{character}]\cr
The package name, defaults to \code{stringi::stri_extract_last_words(usethis::proj_path())}.}

\item{pkg_path}{[\code{character}]\cr
The path to the package directory, defaults to \code{usethis::proj_path()}.}

\item{template}{The template to add to the NEWS file Can be c("first", "develop").
Defaults to NULL.}
}
\value{
None. Creates or updates the NEWS file in the package directory.
}
\description{
Create or update the NEWS file. It will include a template for New section if relevant and update the version number.
}
\details{
Default is  is \code{template = NULL} that will only update the version number.
    Use \code{template = "first"} to make the NEWS template for the first
    release of the package. Use \code{template = "develop"} to make the NEWS
    template for the successive releases of the package.
}
\examples{
\dontrun{
# Create the first NEWS file
library(NVIpackager)
update_news(template = "first")
}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
